Ga entity explorer optimization (#12)
* Changed the sizes for some of the buttons and the currency div on the complience screener.
Added some color to the Register button

* Removed some of the styling not used

* Working on the Entity Explorer
Working on status, centralized? and KYCreq? section
Created an OFAC sanctioned warning under the proper entities

* Modifications to entity status
Removed status if KYC is required
Removed status if entity is active
Removed status if entity is centralized

Added messafe if Entity is Decentralized entity
Added message if No KYC is required
Added message if entity is inactive or does not support crypto
Added a sanctioned tag if the entity is sanctioned by OFAC

* Removed sections where we dont have any information on the Entity

* Some modifications spacing the elements correctly

---------

Co-authored-by: Gerardo Acedo <gerardoacedo@192.168.1.4>
Co-authored-by: Gerardo Acedo <gerardoacedo@192.168.1.19>